+++ kernel/src/Memory/kmalloc.cpp
@@ -22,6 +22,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #include <stdio.h>
 #include <util.h>
 #include <assert.h>
+#include <math.h>
 
 #include <Memory/PageManager.hpp>
 
@@ -247,7 +248,7 @@ void mrvn_memory_init(void* mem, size_t size) {
 }
  
 void* mrvn_malloc(size_t size) {
-    //fprintf(VFS_DEBUG, "%s(%lx)\n", __extension__ __PRETTY_FUNCTION__, size);
+    //dbgprintf("%s(%lx)\n", __extension__ __PRETTY_FUNCTION__, size);
     size = (size + ALIGN - 1) & (~(ALIGN - 1));
  
 	if (size < MIN_SIZE) size = MIN_SIZE;
@@ -283,7 +284,7 @@ void* mrvn_malloc(size_t size) {
 	//printf("AAAA\n");
     mem_free -= size2;
     mem_used += size2 - len - HEADER_SIZE;
-    //fprintf(VFS_DEBUG, "  = %p [%p]\n", chunk->data, chunk);
+    //dbgprintf("  = %lp [%lp]. mem_free = %lx\n", chunk->data, chunk, mem_free);
     return chunk->data;
 }
  
@@ -368,9 +369,9 @@ bool g_kmalloc_initialised;
 void kmalloc_init() {
     g_kmalloc_initialised = false;
     
-    void* pages = WorldOS::g_KPM->AllocatePages(2048);
+    void* pages = WorldOS::g_KPM->AllocatePages(1024); // ~4MiB
     assert(pages != nullptr);
-    mrvn_memory_init(pages, MiB(8));
+    mrvn_memory_init(pages, PAGE_SIZE * 1024);
 
     g_kmalloc_initialised = true;
     NewDeleteInit();
@@ -414,4 +415,54 @@ extern "C" void* krealloc(void* ptr, size_t size) {
     fast_memcpy(ptr2, ptr, old_size);
     kfree(ptr);
     return ptr2;
+}
+
+
+bool g_kmalloc_eternal_initialised = false;
+
+void* g_kmalloc_eternal_mem = nullptr;
+size_t g_kmalloc_eternal_free_mem = 0;
+size_t g_kmalloc_eternal_used_mem = 0;
+
+void kmalloc_eternal_init() {
+    g_kmalloc_eternal_initialised = false;
+
+    void* pages = WorldOS::g_KPM->AllocatePages(512); // ~2MiB
+    if (pages == nullptr)
+        return;
+    g_kmalloc_eternal_mem = pages;
+    g_kmalloc_eternal_free_mem = 512 * PAGE_SIZE;
+    g_kmalloc_eternal_used_mem = 0;
+
+    g_kmalloc_eternal_initialised = true;
+}
+
+extern "C" void* kmalloc_eternal(size_t size) {
+    size = ALIGN_UP(size, 8);
+
+    if (!g_kmalloc_eternal_initialised || size > g_kmalloc_eternal_free_mem)
+        return nullptr;
+
+    void* mem = g_kmalloc_eternal_mem;
+    g_kmalloc_eternal_used_mem += size;
+    g_kmalloc_eternal_free_mem -= size;
+    g_kmalloc_eternal_mem = (void*)((uint64_t)g_kmalloc_eternal_mem + size);
+
+    return mem;
+}
+
+extern "C" void* kcalloc_eternal(size_t num, size_t size) {
+    size = ALIGN_UP(size * num, 8);
+
+    if (!g_kmalloc_eternal_initialised || size > g_kmalloc_eternal_free_mem)
+        return nullptr;
+
+    void* mem = g_kmalloc_eternal_mem;
+    g_kmalloc_eternal_used_mem += size;
+    g_kmalloc_eternal_free_mem -= size;
+    g_kmalloc_eternal_mem = (void*)((uint64_t)g_kmalloc_eternal_mem + size);
+
+    fast_memset(mem, 0, size / 8);
+
+    return mem;
 }